Added support for svn switch
Switch dialog also includes a new repository browser dialog for selecting the target URL.  This can
be used in other actions that need a URL, like svn merge.

Currently, we just run the action and do not do anything to prevent you from shooting yourself in
the foot.  I did not want to do anything that prevented a user that knows what they are doing from
using the option the way that they want.  There are probably some validations that ought to be
added in the future.  The main one I can think of, is that if your run the switch command on a
project, we should probably verify that the target URL has a .project file under it.  Currently, it
does NOT do any of this.

Issue #: 181