New Branch/Tag option
This adds an option to create a branch/tag from the workspace.  The UI once gain mimics TortoiseSVN.
By default, the branch/tag on the server from the HEAD revision of the URL you have loaded, but there
is a check box, you can uncheck to create the branch/tag from your WC.  Presumably, this would 
allow a mixed-revision tag.  This latter option is not currently working.  I think it is because we pass
HEAD for the revision.  From the CLI, the copy command works if the -r parm is not passed.  However,
if -r HEAD is passed, then it works the same as this option does currently.  I am not sure what the
right way to code this to mimic the proper command line behavior.

Also, added a Refresh option to the Choose URL dialog so that you could see your new branch/tag
if you try to switch to it.