Second half of bug-tracking integration.
See: http://svn.collab.net/repos/tortoisesvn/trunk/doc/issuetrackers.txt

Resource history view now turns the bug ID's in commit messages into clickable hyperlinks.  While
making that change, I also made any URL that appears in a commit message into a hyperlink.  So
the above URL should also be clickable.

The one thing we could not do that someday we should, is when showing the Resource history of
an item from the Repository browser, the issue ID's are not clickable.  This is because the current
API's to retrieve properties only works for local items, not URL's.