Improve UI of Switch, Merge and Branch/Tag dialogs.
In all three dialogs, the URL fields have been turned into combo boxes that remember previously
entered values.  The "memory" is on a per-project basis.  So, as an example, if you take the option
to create a Branch/Tag, and then follow that with the Switch option, then the branch you just
created will be in the dop-down.  Likewise, if you are working on a branch and constantly merging
in changes from Trunk, the URL for Trunk will be remembered in the dialog.

In testing this change on OS X, I found what I believe is an SWT bug that made me alter the UI
of the merge dialog.  In that dialog, there are two URL fields, and by default the second URL was
disabled, as it was supposed to be the value of the first URL.  What was happening is that when the
first URL would change, the second URL would not update to the same value.  I think there is a bug
that a disabled combo box will not accept a change, even programmatically.  To work-around this
problem, instead of disabling the second combo, I now hide/show it and when it is hidden, I just use
the value from the first URL in both places.