Added Mark as Merged action to Synch view.  This is different than Mark Resolved which is for a Subversion conflict.
Mark as Merged assumes you have manually merged the local copy of the file.  What it does is set aside the current
file.  It then does a revert and an update, followed by a restore of the edited file.  This causes the WC to be all set
to allow a commit to happen.

Also, modified the Commit action to only show on Outgoing changes, since you cannot commit a change that has a
conflict.  Commit will not work if WC is not up to date (Sync Conflict) or if there is a Subversion conflict.