Enhancements to merge dialog.
Our dialog was modelled on the dialog used in TortoiseSVN.  There were some discussions on
their mailing lists about improving the UI of this dialog.  We are mirroring the changes that
they made to their dialog in ours, and added some new options in the process.

1)  Layout and terminology was tweaked to reinforce the idea that you a merge is taking
the diff between 2 URL/revisions and applying that diff to your WC.

2)  A new Show Log button was added that lets you browse the history of the URL and select
revisions from that dialog.  The "select revision" process tries to figure out what kind of merge
you might be doing and act accordingly to help you.

a)  If you take Show Log from the WC button, and select a revision, it will apply that revision to
the From revision field, provided that it is empty.  The idea here is that you are finding the revision
for the last time you did a merge and want to use that as the starting point for this merge.  This is
the "repeated" merge scenario of keeping a branch up to date with trunk or another branch.

b)  If the "Use From URL" button is checked and you take the Show Log option from either the From
or To revision fields, then the "revision range" you selected will be filled into the two fields.  The from
revision will be the lowest revision number selected minus one (N -1) and the highest revision will be
filled in the To revision.  So if you selected just revision 100, then 99 would be the From revision and
100 would be the To revision.  This scenario is the "port" revisions from one branch to another
scenario.

c)  Finally, if the Use From URL button is not checked, then the revision you select is what is filled in
to the field, and only that field.  

In addition to these changes, buttons were added to a Dry run merge and also to produce a Unified 
Diff file.  These two options allow you to "test the merge" settings.  As such, both options leave the
dialog open so that you can then just press Merge with the same settings.

The Dry run option currently only works for the Command Line adapter, but support for this option
should make it into JavaHL in svn 1.2 timeframe.