Enhancements to commit dialog and resource history view.
A new preference for SVN now exists under Workbench -> Colors and Fonts to specify the font to use for
the commit comments.  It defaults to Text Font, which should be a monospaced font (which is what we
ideally want).

Added support for the following SVN properties that can be set on the Project folder:

(These all come from the same feature in TortoiseSVN)

tsvn:logwidthmarker

If specified, a line will be drawn in the commit comments dialog to show that line number.  Many projects
like Subversion itself, do not want you to go part 80 lines in a single line of comment.  This is an aid to
do that properly.  You must have a moospaced font for this to be 100% accurate.

tsvn:logminsize

If specified, the OK button will not be enabled until the log comment contains at least this many characters.

tsvn:logtemplate

If specified, the commit comment will be initially filled in with the contents of this property.