Modified the icon we use to decorate an SVN conflict to be a warning icon.  The previous icon was the same one used to show a conflict in the Synch view and since they are different kinds of conflicts, that was confusing.
A warning icon is a bit confusing, primarily because they are used for so many
other things in Eclipse.  However, this situation is really a warning/error
condition from Subversion so it make sense that it be a warning or error icon.

When you have a conflict in source code, the Subversion conflict markers cause the
source to be in error, so you typically see the Eclipse error icon.  Therefore, an
error icon for the decorator would be bad.  Likewise, I tried some other red-themed
icons and they do not look good next to the error icon.

The final icon choice was the merged_ov.gif we already include but do not use in
our overlays.  This icon looks good next to an error icon, but I just didn't feel
it was communicating anything about the situation to the user.

So I decided to go with the warning icon.  It can always be changed once people
see it for real.

Finally, one benefit of the warning icon, is that the two conflict-related actions
we provide -- Edit Conflicts and Mark Resolved, both use this symbol in their icon
already.