Fixed NPE when fetching affected paths for an item revision where the HEAD location of the item did not exist at that revision.
Simply changed it to fetch using the repository URL instead of the resource
URL.  Since we are fetching a specific revision, there is no need to provide
a resource URL.  By using the repository URL this should work.  

If the repository root is defined, that is the URL used, otherwise, it uses
the repository URL.  In the latter case, if the URL were to a specific branch
it is possible to still get this problem.  In that case, the solution is
simply specify the correct repository root.

Or wait for Subversion 1.2.



Issue #: 275