Added WorkspacePathValidator class which checks if the case of the Eclipse workspace matches the actual file system case.  When it doesn't (on Windows) it causes problems because svn returns the file system paths, and Eclipse has a different value.  So we are not able to turn the svn paths into eclipse resources.
If the paths do not match, we display a message explaining things to the user,
including how to fix it.

Added call to this class in Checkout, Create Repository and Share Project.  This
should prevent people from getting into a problem situation.

Also entered a bug in Eclipse to have them canoncialize the workspace path so that
this cannot happen.
Issue #: 285