Make the Checkout progress info more verbose so that you can see stuff is happening even if you do not show the console view.
Patch from Martin Letenay

As I've mentioned before I think some long operations are too silent and deserve
more information for the waiting user.  As first victim I choose Checkout.
 
I've noticed that in CheckoutCommand there was a lot of duplicity by having two
versions of run() and scrubProjects() methods.  The two versions were actually
identically, except for the code for monitoring. (And of course the run(monitor)
is also forking runnable interface)  Honestly I hate duplicate code so I merged
these methods back to a single one with the monitor argument as optional.
 
According to log message, these non-monitored versions were created to allow to run
them as a background operation. 
 
So after I've merged them into single methods, I created a clone of "backgroundable"
run() method with a name runDirectly().

This runDirectly() method is now being called from CheckoutAsProjectOperation.
 
To get the progress messages from the svnClient I added an optional notifyLister
to the OperationManager.

At the moment it only listens for the logMessage().

For the convenience reasons, there is also a ISVNNotifyAdapter there ...
 