Institute an official project changelog for Subclipse.
The file is in HTML and will be stored in the feature-plugin since it can apply
to all aspects of Subclipse.

We can link to this HTML from our web site, as well as within our documentation.

After the repository is moved, I am also going to pull this file into the
feature project via svn:externals.  Then, we can also link to this document
with the "More Info" button that exists when a user is installing Subclipse.

Ideally all devs will maintain this file as they commit new changes.