Patch from Martin Letenay
- removed now obsolete StatusCacheComposite
- removed option to save cache status from code and GUI 
- fixed wrong handling of unversioned resources
- fixed the svn:ignore syncView issues
- fixed the right-click delay
- fixed the syncView problem when deleting whole trees of directories

+ implemented getResource() method for LocalResourceStatus so it now
properly determines the resource kind (either from ResourceInfo or
svnstatus.kind)
- removed GetStatusCommand#getResource and replaced it's calls by direct
LocalResourceStatus#getResource()

There are still some issues to be resolved, but most of them are related to
performance, not correctness anymore ...

Just for the record:
- sharing already existing project does not work properly (the sync info is
get before the wizard asks for it and the .svn files are not flagged as
team_private)
- during checkout as and also many other times the "svn status" is invoked
too many times on e.g. sibling resources. Proper use of recursive update
strategy should be considered here.
- there are some error message in SVNWorkspaceSubscriber which are not
translated in the .properites file.
- when directory property change is reverted (e.g. add to svn:ignore), the
child resources (e.g. the ignored file) are not refreshed after revert
- when project is disconnected/unshared to status infos are not removed from
resource#resourceInfo's.
- seems like .svn resources are refreshed too many times, might be a problem
with team_private flags ...
- seems like some listeners are notified unecesarily several times

Besides, the synchronization stuff seems to be relatively stable and correct
now in general.