StatusAndInfoCommand.java
Added try/catch on attempt to retrieve SVNInfo.  The problem case is with
svn:externals.  We do not build a correct URL so the svn info call fails.  
The better fix would be to build the correct URL and let exceptions be thrown
as they were prior to this change.

That being said, this change seems to fix the problem completely and the
extra call to svn info only happens on new incoming files and is not 
absolutely essential for things to work.


Issue #: 343