Fixed the performance problem during checkout operations.
The change consists of 2 parts.
First, the checkout command is now run as an atomic workspace operation, 
so change deltas created during checkout phases are not propagated immediately,
but only after the whole operation is finished. 
(In the old code it used to happen that folders changes were broadcasted first,
and all the files later, which caused 2 times many status updates.)

Second, FileModificationManager was refreshing statuses of all changed resources,
but since we are not able to get/refresh status of single file anyway, 
it was pretty inefficient. Now only folders status are being refreshed (with 
depth one obviously).

For this to happen new run() method was introduced to SVNProviderPlugin, where
one can also specify the scheduling rule (instead of default whole workspace rule).