Support for JavaSVN as a ClientAdapter.  Includes JavaSVN 0.9.1 JAR files.
The previous implementation is being deleted.  This new way is much simpler.  JavaSVN
provides an implementation of the JavaHL SVNClientInterface.  So instead of replacing
JavaHL, it just implements the interface.  I moved most of the code in the existing
JavaHL adapter to an AbstractJhlClientAdapter class.  This serves as the base for
the JavaHL and JavaSVN adapters allowing them to share most of their implementation.

The adapters themselves mostly just have to construct the appropriate implementation
of the interface.

In the JavaSVN adapter I currently have it constructing either a JavaHL or CLI adapter
internally so that it can provide an implementation of the createRepository() method.
We could potentially take advantage of this in the future to support new Subversion
features that JavaSVN does not support, as happened with the Lock options in 1.2.

At the time of this commit, JavaSVN supports all Subversion features, except file://
URL's.