Improvement to Checkout As.. action.
Previously, after the New Project wizard we delete every file that the wizard 
created because we do not want to checkout command to abort if it runs into an
existing file.

Now, before deleting something, we first construct a URL out of the item and run
svn info to see if the item exists in the repository.  We only have to delete
the items that do exist.

Currently, when the item does not exist, an exception is shown in the SVN Console.  
We could possibly change that.