Added new methods to svnClientAdapter, and an implementation for JavaSVN.
commitAcrossWC() - This commit method can accept paths from multiple working
copies as well as a boolean as to whether to perform the commit atomically.
JavaSVN has added support so that when multiple disjointed working copies
all belong to the same repository it can commit all of them in one atomic
operation.  It will figure out if the working copy is for a different
repository and process that in separate commit.

canCommitAcrossWC() returns a boolean that indicates if the previous method
is supported.  Currently, only JavaSVN returns true.