Make the Action column resizable in History View.  This allows someone working in a localized version to see the full text.
Patch by: Takashi Okamoto

I made a minor change to the patch.  Namely, I did not make the initial size
of the column quite as wide as he did.  I imagine this means that someone
working in the Japanese locale would have to resize the column to see the
entire string.
Issue #: 413