Add support for svn 1.3 SVN_ASP_DOT_NET_HACK feature.
NOTE: This feature initially created some hard crashes in JavaHL,
even on Linux.  It wasn't the feature itself, but attempting to call
the new getAdminDir.. isAdminDir.. methods (probably in the middle of
the library doing another action).  By reworking the code to avoid
calling those methods I believe I was able to correct the problem.