More startup optimizations.
These changes have helped with the problem of the 5000ms timeout that can
appear in the log, but it doesn't remove it.  I am not even sure if that is
possible.

These changes makes the plugin start/initialize faster but if it doesn't happen
within 5000ms you can still get the message in the log.