Do not show the username and password fields for a repository connection unless the user is using the CLI adapter.
In order to support existing users, the properties dialog for an existing
connection will still show these fields if they are not already null or empty.

The reason for this is that it is better to allow JavaHL or JavaSVN to ask
the user for this information when/if it needs it.  It also eliminates some of
the confusion around SSH connections where the username and password have to
be specified elsewhere anyway.