Preliminary support for Issue#: 74.
In the MoveDeleteHook when a file is deleted we will not look
for the presence of a property named "DeferFileDelete".  If it
exists with a value of true, then we will ignore the delete and
not run svn delete.  That way, if a generator is just deleting
the file and replacing it with a new version we will pick it up as
modified.

Still need to do some more work to better handle the case where the
file is not put back and the Subversion marks it as missing.
Issue #: 74