Enhancements to Pending Operations view:
1)  Shows files with a "missing" status.

2)  When files are missing there is a "Mark As Deleted" action
on the Team menu.  This will let you run svn delete so that its
status is correct in the working copy.  Of course the Revert
action also exists to put the file back.

3)  When the view is opened or refreshed it runs an Eclipse
refresh on the selected item.  This is to make sure that the
proper status of all items is known.  For example, if an item
is missing in the file system, but Eclipse does not know it yet,
you could get weird results.