Refactoring of three checkout options into a single checkout wizard similar to the one that CVS uses.
Also added a New Project from SVN and Import Project from SVN options that run the
wizard.

The only CVS feature not currently supported is checking into existing project.
The checkout part of that is probably not too difficult, it would be making
all of the other parts work that is hard.

There is a problem in this wizard when creating a new location as part of the wizard.
I want to get this committed and work on it some more.