Second half of the change for issue #402.
In the commit UI do not allow the user to run the commit if they have a folder
delete and a folder prop-change selected but not all of the files beneath
the folder prop-change are selected.

This is because a folder delete requires a recursive commit, and to commit a
folder prop-change without it's sub-content requires a non-recursive commit.
Issue #: 402