Added a Show Properties option to the Synchronize view.  Enabled on outgoing items only.
Also, fixed the selection processing on Revert so that it processes the specific
item you select (including folders) rather than the items beneath it.  I found
that unversioned items could show in the dialog which was incorrect.  If you select
a folder the option should work the same way it would if you did it from the 
Navigator view.