Fixed handling of commit from Synch view.
It is now based on the actual selections in the view, and it still does them
all as a single commit if using JavaSVN.  The dialog only comes up once
regardless.

If multiple ChangeSets are selected, the commit messages are combined (as they
are in CVS), and the ChangeSet command comes up even if you do not select
the changeset itself.
Issue #: 506