Fixed problem with Compare with Latest when there were spaces in the filename.
To do this correctly, I needed to add a new method to ISVNClientAdapter that
exposes the revision parameter on the svn info command.  Then changed the code
to use that method.  I also changed one other place in the code that could
be improved by using this new method.

Issue #: 512