Change update from the Synch view to operate on the items in the view as opposed to the selection.  This is how it used to work.
Huh?

The "folders" in the Synch view are put there by the view based on the resources
you show in the view.  They are not considered items in the view unless those
folders happen to have changed.  Such as a property change, or the new change
that will show out of date folders.

Prior to this change, if you selected a folder, and did update, we would run
update on that folder.  Now, we will run it on the items in the folder.  Again,
unless the folder itself is changed.

This makes update more accurate in general, with a possible sacrifice in speed.
Ideally, this would be used in conjunction with the feature that shows out of
date folders, since that will cause them to be "in the view" anyway.

I also included a performance optimization to detect when folders are being
updated, and to not include their children in the call to the command.