Remove code that should not be needed anymore.
The status operation now returns all of the info we need, so there is no
longer a need for those other methods to be called to get the extra info.
Most of the time, those methods just short-circuit return but sometimes items
fall into the methods, and when they do it really is not needed.