Added new preference to "Show Unversioned Resources in Commit Dialog".
When this is not selected, we do not traverse into unversioned folders to
add them to the dialog.  When it is selected, then the previous preference of
"Select Unversioned Resource .." still comes into play.  That preference is
now hidden if the other one is not selected.

Finally, also enhanced this so that when the dialog is brought up as part of
the Share Project process the prefs are ignored and the files are selected 
by default.
Issue #: 508