svn:externals folder is showing as outgoing change.
This change fixes the problem, but it is hard to understand why the current
code was there.  So will have to see if any side effects turn up.  My guess
is that the problems that code was solving have been resolved in our better
gathering of status information for externals.
Issue #: 562