Think this gets change sets working.
The main change is probably the way the extension is declared in the plugin.xml.
I added in the second element that Mylar has for the CVS extension.

I also added a new class and moved code to it for the CommitWorkflow.  As near
as I can tell, this code will not be used anymore as the Commit Context action
was removed from Mylar.  Just making things match what the CVS integration does.

It does seem to work now though.