Support atomic commits across multiple projects when using JavaHL.
This works by exploiting a hack in Subversion.  So this was made a preference
that is off by default.

Basically, if the common parent folder of all the projects contains a .svn
metadata folder, then Subversion will do the commit.  So we create/delete this
on the fly, and do the proper grouping by repository.

A nice side-effect is that if the projects really do have a common parent, and
were checked out using another client and then imported into Eclipse (which
a lot of people do) then this really is not a hack and is just working better.
The hack is the creation/removal of the fake .svn folder so that the commit
will work.
Issue #: 527