Remove calls to svnResource.isManaged() when retrieving ProjectProperties. The process of answering this question can trigger Eclipse to kick off a big refresh process if the resources are out of synch with the file system.
A side effect of this change is that if you are using multi-level 
svn:externals:

project/path/src url://repos/project/src

Then when we try to get the props of path a message will be logged in the console
Everything still works correctly though.  We can probably try to hide the
message if this becomes a problem.

This should hopefully resolve: 
https://bugs.eclipse.org/bugs/show_bug.cgi?id=200220