Attempt to fix more of the random crashes some users get.  Basically, I am reducing the number of times we need to change the JavaHL context (config dir, user prompt, listeners etc.)  When these changes are made in threading situations it can lead to crashes.
So this change streamlined the order things are done at startup, making sure
all preferences are set before we initialize a client.  Then we also stopped
setting the context every time we ask for a client.

The crashes are hard to reproduce so it is unclear if this will resolve the
issue.