Issue 923: Unable to rename/refactor java file with svn:keywords
Override the JavaHL move method to check if a file is being moved, and if so,
then checks if file has svn:keywords and if so, instead of doing svn move, we
do this:

svn copy from to
overwrite contents of to with from
svn delete from

This is equivalent to an svn move and overrides the new SVN behavior of
updating the keywords.  SVNKit would still just do svn move.