Just a hack for now, but refresh statuses recursively.
The issue is that with 1.7 we no longer have .svn folders everywhere.  Our
current logic for refreshing caches is to monitor for changes to these folders
as a trigger.  Now, there is just one of these folders at root of our working
copy.  We really should not be monitoring these as all.  It is possible a user
will not even have one of these in their working copy.  We need a new way
to monitor for changes and refresh cache.